Converted SuperTextField into a platform adaptive widget (Resolves #413)
