+++ super_editor/example/lib/demos/supertextfield/_emojis_demo.dart
@@ -18,7 +18,7 @@ class EmojisTextFieldDemo extends StatefulWidget {
 class _EmojisTextFieldDemoState extends State<EmojisTextFieldDemo> with TickerProviderStateMixin {
   final _textFieldController = AttributedTextEditingController();
 
-  GlobalKey<SuperTextFieldState>? _textKey;
+  GlobalKey<SuperDesktopTextFieldState>? _textKey;
   late TextFieldDemoRobot _demoRobot;
 
   FocusNode? _focusNode;
@@ -95,7 +95,7 @@ class _EmojisTextFieldDemoState extends State<EmojisTextFieldDemo> with TickerPr
                 },
                 child: SizedBox(
                   width: double.infinity,
-                  child: SuperTextField(
+                  child: SuperDesktopTextField(
                     key: _textKey,
                     textController: _textFieldController,
                     focusNode: _focusNode,
