+++ super_editor/example/lib/demos/supertextfield/_expanding_multi_line_demo.dart
@@ -23,7 +23,7 @@ class _ExpandingMultiLineTextFieldDemoState extends State<ExpandingMultiLineText
         ),
   );
 
-  GlobalKey<SuperTextFieldState>? _textKey;
+  GlobalKey<SuperDesktopTextFieldState>? _textKey;
   late TextFieldDemoRobot _demoRobot;
 
   FocusNode? _focusNode;
@@ -89,7 +89,7 @@ class _ExpandingMultiLineTextFieldDemoState extends State<ExpandingMultiLineText
                 },
                 child: SizedBox(
                   width: double.infinity,
-                  child: SuperTextField(
+                  child: SuperDesktopTextField(
                     key: _textKey,
                     textController: _textFieldController,
                     focusNode: _focusNode,
