+++ super_editor/example/lib/demos/supertextfield/_interactive_demo.dart
@@ -132,7 +132,7 @@ class _InteractiveTextFieldDemoState extends State<InteractiveTextFieldDemo> {
             },
             child: SizedBox(
               width: double.infinity,
-              child: SuperTextField(
+              child: SuperDesktopTextField(
                 textController: _textFieldController,
                 focusNode: _focusNode,
                 textStyleBuilder: _textStyleBuilder,
