+++ super_editor/example/lib/demos/supertextfield/_single_line_demo.dart
@@ -22,7 +22,7 @@ class _SingleLineTextFieldDemoState extends State<SingleLineTextFieldDemo> with 
         ),
   );
 
-  GlobalKey<SuperTextFieldState>? _textKey;
+  GlobalKey<SuperDesktopTextFieldState>? _textKey;
   late TextFieldDemoRobot _demoRobot;
 
   FocusNode? _focusNode;
@@ -84,7 +84,7 @@ class _SingleLineTextFieldDemoState extends State<SingleLineTextFieldDemo> with 
                 },
                 child: SizedBox(
                   width: double.infinity,
-                  child: SuperTextField(
+                  child: SuperDesktopTextField(
                     key: _textKey,
                     textController: _textFieldController,
                     focusNode: _focusNode,
